breakout logic functions into their own files
