+++ client/src/routes/app/[country]/[map]/[year]/+page.server.ts
