+++ client/src/routes/app/[country]/[map]/[year]/initialize.ts
@@ -54,10 +54,11 @@ function initializeMap(mapBind: HTMLDivElement, candidates: Candidate[]) {
 					(text as HTMLElement).style.color = luma > 0.5 ? 'black' : 'white';
 				}
 
-				if (childHTML.hasAttribute("disabled")) { //Set disabled styles & remove state value, styles from yapms1
+				if (childHTML.hasAttribute('disabled')) {
+					//Set disabled styles & remove state value, styles from yapms1
 					candidate.margins[0].count -= value;
-					childHTML.style.fillOpacity = "0.25";
-					childHTML.style.strokeOpacity = "0.25";
+					childHTML.style.fillOpacity = '0.25';
+					childHTML.style.strokeOpacity = '0.25';
 				}
 			}
 			if (texts) {
